Добавлена новая реализация расположения класов. Теперь окна расположены в отдельном файле. Добавил файл админа, карточки фильма, а также класс переопределенного списка.
